requirements.txt
pytorch
pytorch_lightning
sklearn
numpy
rdkit
networkx
